backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-cors-headers==3.10.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==3.3
install==1.3.4
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
pytz==2021.3
requests==2.26.0
simplejson==3.17.5
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
